Close #262 - Support Scala 3.0.0
