+++ .github/workflows/build-graal.yml
@@ -10,8 +10,8 @@ on:
       - main
 
 env:
-  CLI_SCALA_VERSION: "3.0.0-RC3"
-  CLI_SCALA_BINARY_VERSION: "3.0.0-RC3"
+  CLI_SCALA_VERSION: "3.0.0"
+  CLI_SCALA_BINARY_VERSION: "3"
   GRAALVM_JAVA_VERSION: "adopt@1.11"
   GRAALVM_BIN_DIR_NAME: "native-image"
