--- .github/workflows/manual-github-upload-artifacts-all.yml
@@ -1,43 +1,0 @@
-name: "[M] GitHub Release: Upload Artifacts"
-
-on:
-  workflow_dispatch:
-
-jobs:
-
-  github-release-artifacts-and-publish:
-
-    runs-on: ubuntu-latest
-
-    strategy:
-      matrix:
-        scala:
-          - { name: "Scala 2", version: "2.12.12",   binary-version: "2.12",      java-version: "11", cli: "" }
-          - { name: "Scala 2", version: "2.13.3",    binary-version: "2.13",      java-version: "11", cli: "" }
-          - { name: "Scala 3", version: "3.0.0-RC1", binary-version: "3.0.0-RC1", java-version: "11", cli: "cli" }
-
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-java@v1
-        with:
-          java-version: ${{ matrix.scala.java-version }}
-
-      - name: Cache SBT
-        uses: actions/cache@v2
-        with:
-          path: |
-            ~/.ivy2/cache
-            ~/.cache/coursier
-            ~/.sbt
-          key: ${{ runner.os }}-sbt-${{ matrix.scala.binary-version }}-${{ hashFiles('**/*.sbt') }}-${{ hashFiles('**/build.properties') }}
-          restore-keys: |
-            ${{ runner.os }}-sbt-${{ matrix.scala.binary-version }}-
-
-      - name: "sbt GitHub Release and Publish tag for ${{ matrix.scala.name }} ${{ matrix.scala.version }} - ${{ github.run_number }}"
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_AUTH_TOKEN }}
-          BINTRAY_USER: ${{ secrets.BINTRAY_USER }}
-          BINTRAY_PASS: ${{ secrets.BINTRAY_PASS }}
-        run: |
-          export GITHUB_TAG=${GITHUB_REF#refs/tags/}
-          .github/workflows/sbt-github-upload-artifacts.sh ${{ matrix.scala.version }} ${{ matrix.scala.cli }}
