+++ .github/workflows/release.yml
@@ -9,8 +9,8 @@ on:
 
 env:
   GH_JAVA_VERSION: "adopt@1.11"
-  CLI_SCALA_VERSION: "3.0.0-RC3"
-  CLI_SCALA_BINARY_VERSION: "3.0.0-RC3"
+  CLI_SCALA_VERSION: "3.0.0"
+  CLI_SCALA_BINARY_VERSION: "3"
   GRAALVM_JAVA_VERSION: "adopt@1.11"
   GRAALVM_BIN_DIR_NAME: "native-image"
