+++ core/src/main/scala-2.12_2.13/maven2sbt/core/Libs.scala
@@ -1,6 +1,5 @@
 package maven2sbt.core
 
-import scala.language.implicitConversions
 
 import io.estatico.newtype.macros.newtype
