+++ core/src/main/scala-2.12_2.13/maven2sbt/core/core.scala
@@ -5,7 +5,6 @@ import io.estatico.newtype.macros._
 import io.estatico.newtype.ops._
 import just.fp.Named
 
-import scala.language.implicitConversions
 
 /**
   * @author Kevin Lee
@@ -32,9 +31,13 @@ package object core {
     def unapply(groupId: GroupId): Some[String] =
       Some(groupId.coerce)
 
+    implicit final class GroupIdOps(val groupId: GroupId) extends AnyVal {
+      def render(propsName: Props.PropsName): RenderedString =
+        GroupId.render(propsName, groupId)
+    }
+
     def render(propsName: Props.PropsName, groupId: GroupId): RenderedString =
       StringUtils.renderWithProps(propsName, groupId.value)
-
   }
 
   @newtype case class ArtifactId(value: String)
