+++ core/src/main/scala-3/maven2sbt/core/MavenProperty.scala
