+++ core/src/main/scala-3/maven2sbt/core/Props.scala
@@ -4,7 +4,6 @@ package maven2sbt.core
  * @author Kevin Lee
  * @since 2020-12-13
  */
-@SuppressWarnings(Array("org.wartremover.warts.ImplicitConversion", "org.wartremover.warts.ImplicitParameter"))
 object Props extends PropsPlus {
   opaque type PropsName = String
   object PropsName {
